requirements.txt
fastapi
uvicorn
pydantic
pytest
python-dateutil
